Refactor [#13] Task에 해당하는 리뷰사항 관련 수정 작업 및 패키지 관련 변경 작업 진행 + Fix conflict feat#9 #15 resolve
- 구현 이슈 관련하여 추가 수정 진행 user -> category -> task 가 아닌 바로 task로 가서 찾는 방법으로 진행
